WatermarkService/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
torch==2.4.0
torchvision==0.19.0
pillow==10.2.0
numpy==1.26.3
videoseal @ git+https://github.com/facebookresearch/videoseal.git
# C2PA Content Credentials
c2pa-python>=0.6.0
cryptography>=42.0.0
